refactor: refactor fs task

+++ src/fs/delete.js
@@ -1,14 +1,21 @@
 import { unlink } from "fs/promises";
-import { dirname, join } from "path";
-import { fileURLToPath } from "url";
+import { join } from "path";
+
+import { exists, pathToDir } from "./shared.js";
 
 export const remove = async () => {
-  const __dirname = dirname(fileURLToPath(import.meta.url));
+  const __dirname = pathToDir(import.meta.url);
   const pathToFile = join(__dirname, "files/fileToRemove.txt");
 
   try {
+    const isPathToFile = await exists(pathToFile);
+
+    if (!isPathToFile) {
+      throw new Error("FS operation failed");
+    }
+
     await unlink(pathToFile);
-  } catch {
-    console.log(new Error("FS operation failed"));
+  } catch (error) {
+    console.error(error.message);
   }
 };
